ci: move state to s3
