+++ package.json
@@ -42,6 +42,7 @@
     "node": ">=22.11.0"
   },
   "dependencies": {
+    "@cdktf/provider-aws": "^19.40.0",
     "@cdktf/provider-github": "^14.3.1",
     "cdktf": "^0.20.9",
     "constructs": "^10.4.2",
